+++ src/components/sections/SentimentSection.jsx
@@ -1,5 +1,6 @@
 import React from 'react'
 
+// eslint-disable-next-line no-unused-vars
 const SentimentSection = ({ closeSection }) => {
   return (
     <article id="sentiment" className="max-w-4xl">